fix: Throw error when diff uploading >2mb
As of writing this, vexide's differential upload patcher cannot handle files greater than 2mb in size (this includes the patch, base, and new binary). Differential uploading binaries greater than 2mb is undefined behavior, so this is not something that should ever silently fail. This commit adds errors to prevent uploading files this large.